move: scaffolding to watch upgrade objects
## Description 

Stacked on https://github.com/GeunhwaJeong/haneul/pull/13376.

Adds a function `watch_for_upgrades` to track upgraded packages by
watching Object IDs. The function is more of a scaffolding to get tests
in place for testing the callback on `haneulx_subscribeTransaction`. So it
is not complete and has the following limitations (to address in follow
up):

- The function is not used in the main event loop of the server yet, it
is only an entry point for tests

- The function returns after observing one event / transaction, to
simplify test set up. In follow up I will put it in a loop and figure
out an elegant way to test it as a long running process (I think I'll
likely need to pass a callback to the function).

- Only creates a client for `localnet` right now. This will be updated
so that there is a client for each network of watched packages in
`config`.

## Test Plan 

Added test to set up the watch function, upgrade a package, and expect
to observe an event.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes