[move-compiler] A fix to the freeze wrapped linter
## Description 

The original implementation of the freeze wrapped linter
(https://github.com/GeunhwaJeong/haneul/pull/13194) had a bug - arguments to
the transfer function were not processed to look for other cases where
the linter warning should kick in. The case would have to look similarly
to the example below, which makes it pretty unlikely, but still...

```
    public fun freeze_arg(w1: Wrapper, w2: Wrapper) {
        transfer::public_freeze_object({ transfer::public_freeze_object(w1) ; w2});
    }
```

## Test Plan 

Added a test capturing the missing bit (same as example above)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes