Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/deepbook@0.4.0

### Minor Changes

-   93055a2bb: Migrate to CLOB V2

## haneullabs/zklogin@0.1.1

### Patch Changes

-   c003636c0: Change function signature of genAddrSeed
-   d3d1d94b7: Fix generated types
-   a4fe5d5c3: Update max key claim value length
-   dd5220a92: Initial experimental zklogin SDK
-   2102be90a: Polish utils
-   9476e6e73: Change build process to use new internal build process.