[1/2] Move TempStore and GasCharger to execution layer
## Description 

This PR makes it such that we always construct TempStore and GasCharger
inside execution layer, and there is no more access to either of these
two types outside of execution. This sets up the code in the right shape
for step 2 of actually moving them.

## Test Plan 

Non functional change, CI.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes