Properly mark deleted child object as modified
## Description 

Explained in the comment. Some child objects are deleted without having
an entry in the child object effects. We need to account for them,
otherwise we may incorrectly mark deleted object as
unwrapped_then_deleted.

## Test Plan 

I need to look into how to write PT to uncover this. Fixing this asap to
make sure no bugs are shipped.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes