Do not read DB for storage rebate collection
## Description 

When collecting storage rebate of input objects for mutations, currently
we read from DB for each child object. This is inefficient and
unnecessary.
Change it to call a helper function that we already have to read from
in-memory metadata instead.
This conflicts with https://github.com/GeunhwaJeong/haneul/pull/13519 on this
function as they both got rid of the extra writes. dariorussi let me
know whether you would prefer to take this into your PR.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes