Remove session
## Description 

This is the first step in removing Session from the execution
layer/adapter.
It's a replacement for  as it creates a new version of the
execution layer.
It also adds a couple of protocol config features around publishing, a
change in charging and in the number of packages published per PTB.
Here we get rid of Session entirely. However there is a lot more we are
doing in the VM (loader and more) and in the adapter traits. More
coming.

I have the change for protocol config here, however I could remove it if
we want to go in without the protocol change yet. Looking for feedback
around how we want to push this.
We can enable this only for devnet (maybe testnet) if we wanted, or go
to mainnet, or nowhere for now.

tnowacki amnn sorry for the churn but this should be what you reviewed
already. As a side note I went back to have the published packages as a
vector, given the new constraints. To be honest I feel that logic will
change more as we integrate better with the Move VM, and that is coming
soon. Not sure if that is a reason to keep this on the side, or to
promote it sooner because of the new protocol changes

## Test Plan 

More tests added for new features

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes