[dapp-kit] create new wallet provider for managing wallet connection state
## Description 

This PR is the first part in a series of work to add wallet connection
management functionality to `dapp-kit`. The general plan is that we are
going to use native state management in React combined with
`react-query` to expose a new series of hooks that will replace
`wallet-kit`.

For this PR specifically, we're scaffolding out the initial context,
state management, and new hook called `useWallet` that exposes some of
the wallet connection state such as accounts and wallets. To answer a
few potential questions on this one:

Why are we using a reducer for state management vs. normal `useState`,
zustand, etc.?
-> Most/all of the wallet state is intertwined enough that it makes
sense to use a reducer over `useState`. Since we're going for simplicity
with the opinionated React architecture, we're going to roll with this
for now over something like Zustand.

Why are we adding a `connectionStatus` state instead of relying on
mutation state from `useConnectWallet`?
-> Mutation state can't be shared in react-query v4, so we need to
manage this state manually
(https://github.com/TanStack/query/issues/2304)

Why no tests?
-> These will be important to have, but setting up useful tests given
the nature of this functionality is a little tricky. Depending on the
resourcing for this, I'll work on some combination of core logic tests
using a mock wallet, some dApp we spin up for E2E tests, or something
along those lines.

## Test Plan
- I have an example Vite application that will be added in a later PR
that I've been testing with
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes