Enable clippy for `external-crates/move` in CI, with clippy changes in the code.
## Description 

Enables clippy for `external-crates/move`, including fixes that clippy
flagged (with a few ignores).

Hopefully this will avoid issues in the codebase in the future that
clippy will flag automatically.

**Most of these changes were made automatically with `cargo clippy
--fix` so special care should be taken in reviewing this diff.**

## Test Plan 

CI should work as expected, and now report signals for `move-clippy`, as
I've called it. (I'm happy to bikeshed the name.)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------