[Deepbook] Adding check for deepbook bid/ asks
## Description 

Originally, deepbook's `getLevel2_ask/bid` functions returned items even
if its out of range. This PR adds a check so that if we try to search
out of range we return the expected empty vectors.

## Test Plan 

How did you test the new or updated feature?

Unit testing

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes