simulacrum: introduce a tool for simulating Haneul chain state
This patch adds a new crate `simulacrum` which contains tooling for
creating an instance of a Haneul Blockchain that is fully simulated. In
other words it is a chain that only advances state when explicitly
cranked. The intent of this is to enable producing fully deterministic
chain states for testing of downstream applications or for smart
contract developers that want a cheap, local development environment
without needing to spin up full consensus.