[GraphQL/Config] TOML-based config
## Description

Introduce config for functional groups to enable and disable groups of
features in the schema, so that different operators can choose to run
the RPC service with different sets of features enabled (for example
running public nodes without analytics).

The precise set of functional groups may shift -- the current list is
derived from the list in the RFC.

This PR only introduces the config, and not the logic in the schema to
listen to the functional group config. The config is read from a TOML
file whose path is passed as a command-line parameter. It is stored as
data in the schema's context so that it can later be accessed by
whatever system limits access to endpoints by flags.

A stub "experiments" section has also been added as a place to keep
ad-hoc experimental flags (to gate in-progress features).

This PR also brings the `ServerConfig` (renamed to `ConnectionConfig`)
into the same module, and applies a common pattern to ensure there's a
single source of truth for default values (the `Default` impl for the
config structs).


Finally, `max_query_depth` is moved from `ConnectionConfig` to
`ServiceConfig` as it's a config that we would want to share between
multiple instances of the RPC service in the fleet.

## Stack

- 

## Test Plan

New unit tests:

```
haneul-graphql-rpc$ cargo nextest run
haneul-graphql-rpc$ cargo run
```