[graphql][2/n] Refactor convert_to_transaction_block
## Description 

Continuing off https://github.com/GeunhwaJeong/haneul/pull/13714. 
It looks like the effects field of TransactionBlock have some heavy
overlaps with other top-level fields of TransactionBlock. We already
have the data needed to resolve every field on effects:
TransactionBlockEffects. In fact, if a caller specifies both
tx.effects.epoch and tx.expiration, we make redundant calls for
haneul_system_state and fetch_protocol_config. It's also a little strange
to me that we pass tx_effects to TransactionBlockEffects for status and
errors. Wondering if it's worth it right now to modify schema for
TransactionBlock and TransactionBlockEffects? We could remove the
redundant digest, lift status and errors to top-level, etc.

## Test Plan 

Some manual queries, updated snapshot tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes