[narwhal] hardening synchronizer gc testing
## Description 

Adding test for the gc logic of the synchronizer State. I've
specifically tested the `State` as it is more straightforward to test
this directly there instead of the Synchronizer it. self as it would be
harder to understand/make it obvious the gc limit.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes