[indexer-v2 1/n] add schemas and typesadd schemas and types
## Description 

Breaking [this](https://github.com/GeunhwaJeong/haneul/pull/13574) into
smaller pieces.
This is first PR to define schemas and basic data models. Addressed most
comments in 


Stack:
1. https://github.com/GeunhwaJeong/haneul/pull/13760
2. https://github.com/GeunhwaJeong/haneul/pull/13774
3. https://github.com/GeunhwaJeong/haneul/pull/13775
4. https://github.com/GeunhwaJeong/haneul/pull/13778

## Test Plan 

Testing is done in a live indexer with the big PR.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
[indexer-v2 1/n] add schemas and types