[EffectsV2 1/n] Simplify inner temp store
## Description 

This PR simplifies the fields of InnerTemporaryStore such that it's no
longer dependent on the implementation details of effects.
Specifically, it gets rid of all the use of WriteKind and DeleteKind
stuff. Instead it only has a map of all written objects.
To access to deleted objects, we could always refer to effects because
we always have effects when we have the inner.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes