[graphql] retire tx digest scalar type
## Description 
In https://github.com/GeunhwaJeong/haneul/pull/13718 there was a request to
remove the dedicated scalar for TransactionDigest, and instead return
digests as String based on the graphql schema and to limit use of custom
scalars. This PR addresses the request, but keeps the Digest type around
to minimize String conversions.

## Test Plan 

snapshot + some manual queries

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes