[dapp-kit] remove WalletAlreadyConnectedError from useConnectWallet
## Description 
Jordan-Haneullabs brought up a good point during a discussion last week
that we shouldn't impose the constraint for people to disconnect a
wallet connection before connecting a new wallet since disconnecting is
really a facade sort of concept. This PR removes the associated logic,
so people can technically connect as many wallets as they please 🔌

## Test Plan 
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes