[dapp-kit] add new useSignAndExecuteTransactionBlock hook
## Description 
This PR adds the final Haneul feature mutation hook for signing and
executing a transaction block. This is similar to
https://github.com/GeunhwaJeong/haneul/pull/13741 and
https://github.com/GeunhwaJeong/haneul/pull/13720.

## Test Plan 
- Automated tests
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes