Extend execution_layer.py to support dashes
## Description 

I wanted to create a feature branch vm-rework, but the script today will
create Rust files and refer to the feature as "vm-rework", which breaks
Rust convention. This PR extends the script so that we can still
reference feature cuts w/ dashes, but will create Rust files and
references with "vm_rework". Verified this works with
./scripts/execution_layer.py cut vm-rework, and then rebasing with
./scripts/execution_layer.py rebase latest

## Test Plan 

Verified this works with ./scripts/execution_layer.py cut vm-rework, and
then rebasing with ./scripts/execution_layer.py rebase latest

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes