[Indexer-v2 4/n] add caches for indexing
## Description 

This PR adds a few things:
(Note that because of the way I'm splitting PRs, some
structures/functions are not immediately in use, but i hope they more or
less make sense).

1. `InMemPackageCache`: an in-mem cache for packages. Because we batch
process checkpoints, it's possible that when we need a module it's not
yet persisted in db. And here's where we can find it.
2. `InMemObjectCache`: an in-mem cache for objects for processing like
object changes or balance changes. Its lifetime is per checkpoint.
3. `TxChangesProcessor`: a helper that leverages `InMemObjectCache` and
implements `ObjectProvider` for dynamic fields processing and others.
4. `InterimModuleResolver` consists of a backup ModuleResolver (e.g.
IndexerStoreModuleResolver) and an in-mem package cache.

Stack:
1. https://github.com/GeunhwaJeong/haneul/pull/13760
2. https://github.com/GeunhwaJeong/haneul/pull/13774
3. https://github.com/GeunhwaJeong/haneul/pull/13804
4. https://github.com/GeunhwaJeong/haneul/pull/13805
5. https://github.com/GeunhwaJeong/haneul/pull/13832
6. https://github.com/GeunhwaJeong/haneul/pull/13834


## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes