[GraphQL/FunctionalGroups] Configuration disables queries
##  Description

Adds an extension to intercept field resolution to check whether the
field being queried belongs to a disabled functional group. If so, an
error is produced instead of allowing the query through.

Changes the functional group enum itself:

- Removing "Packages" (there are no package-specific queries).
- Renaing "NameServer" to "NameService" for consistency.

Currently, disabled fields are not removed from the schema during
introspection. This has been left as a follow-up (TODO in the codebase).

Also, the error from the extension does not supply a source location in
the query because this is not supplied to the extension by
`async-graphql`. Fixing this requires an upstream change, so this has
also been left as a follow-up.

## Test Plan

New tests for extension:

```
haneul-graphql-rpc$ cargo nextest run
```

- All configured functional groups must refer to real fields in the
schema, unless marked as unimplemented.
- All unimplemented fields must *not* exist in the schema (this is to
make sure the tests get cleaned up).
- Querying an enabled field makes it through to the data loader (which
panics, but that's okay).
- Querying a disabled field yields an error.