Add file name random suffix
## Description 

When writing to a bucket which is used for snowpipe ingestion, we need
to randomize the file names. This is because if for some reason we want
to re-run the data pipeline, we will be generating the same file names
in the following runs and writing to the bucket and replacing existing
files. But snowpipe today would not pick up modified files.
## Test Plan 

Existing tests, running locally