[explorer] add missing dependency from previous PR that I accidentally forgot about
## Description 

I don't know how this passed CI, but I forgot to add
`vanilla-extract/vite-plugin` as a dev dependency to Haneul Explorer which
is used for bundling the CSS from `dapp-kit`. This was the original PR:
https://github.com/GeunhwaJeong/haneul/pull/13845

## Test Plan 
- No more local type error in `vite.config.ts`

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes