module bcs data transform to json
## Description 

This PR introduces a new data transformer in `crates/data-transform`

The purpose of this code is to take a row from the `EVENTS` table, read
the bcs coded column, `EVENT_BCS`, and return the decoded JSON, using
the appropriate module.

There are times when the correct package and the module aren't available
in the database, so this uses a custom local resolver,
`GrootModuleResolver`. For modules that do exist, it will use the normal
logic.

The logic is as follows:
-Create a new local module cache with a custom resolver
-Load each row in the events table
-If it is a specific module, use the custom resolver to get the module
from the local cache. Otherwise, use the normal logic
-Deserialize it and return the json
-Write the json to a new table with a json specific column, `EVENT_JSON`

## Test Plan 

How did you test the new or updated feature?

This was developed and tested on my devserver. 
It is currently running and loading results into a Snowflake instance.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------