[dapp-kit] expose the ConnectModal component for creating completely custom buttons
## Description 

This PR exposes the `ConnectModal` component in case folks want to use a
custom button but keep the modal as-is. I added support for controlled
and uncontrolled modes just for additional flexibility (might be
overkill...?).

We're intentionally not going to expose an imperative way to open the
modal sans trigger since it requires some extra work that I don't think
is quite worth it at this time
(https://github.com/radix-ui/primitives/discussions/1234). We can always
re-visit this and add it as a follow-up

## Test Plan 
- Tested in controlled and uncontrolled modes
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes