[move-compiler] revise CFG creation, optimize known jumps
This revises the CFG basic block translation to use the classic
bottom-up style. The code is a bit simpler now.

In addition, this introduces an optimization to forward known jumps: if
a label contains a single jump, this optimization forwards all jumps to
that label to the target label instead.

The tests still work, producing the same results.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

Code compiled with this optimization will produce a slightly-different,
more-concise, structure.