[move-compiler] Make addresses in error messages more readable
## Description 

- About two years ago, we made named addresses transparent, instead of
opaque, with respect to their numerical value.
- To help prevent confusion, the compiler always expanded the named
address.
- Now, the compiler will only do so when the named address is not in a
1:1 mapping with its value

## Test Plan 

- Added tests, updated tests 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

The Move compiler will now try to make addresses in errors more
readable, expanding them to their numerical value only when it is
potentially ambiguous.