Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/dapp-kit@0.2.0

### Minor Changes

-   d32ef851a: Add theme definitions for our UI components
- fd2d0cdca: Theme UI components based on provided theme and add design
polish
- 94194fdbd: Exclude non-Haneul accounts from the accounts state when
someone connects a multi-chain wallet

## haneullabs/kiosk@0.7.2

### Patch Changes

-   5c144882c: Fixes resolve royalty rule from breaking.