move: remove mutually exclusive features
In preparation for merging the move and haneul workspaces, this eliminates
mutually exclusive features from a number of of the move crates. (eg
the ability to specify the length of a move address). This is in order
to enable the ability to easily perform tests, as is done today in haneul,
by passing the `--all-features` to various cargo build commands.