Clean up ConsensusHandler
## Description 

This PR does a few cleanups for ConsensusHandler:
1. Move all the code that are consensus-independent into the function
inside epoch_store. This allows us to get rid of this strange
CommitBatch thing, as well as opens up an interface for other code to
call (e.g. if we want to benchmark the system with a dummy consensus)
2. Move the tx oredering code into its own file for clarity
3. Insert roots at the end to reduce dup entries.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes