move: improve source service error handling
## Description 

Causes the server to shutdown if we see a fatal error when watching for
on-chain upgrades (e.g., a fatal error such as losing a subscription
connection).

## Test Plan 

Tested manually since it's difficult to test forcing a connection drop
to observe the server crashes as expected. I ran a local network and
killed it to see the source server also die. In follow up PR there is
additional testing for the functionality we care about most:
invalidating sources if we observe a upgrade transaction in the
network..

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes