move: source service invalidates data on upgrade txn
## Description 

This PR ensures that as soon as we observe an upgrade transaction, the `sources` reported are cleared and invalidated. It is the conservative approach to ensure we don't report outdated sources.

## Test Plan 

Added test to validate (a) successful upgrade transaction triggers in
`watch_for_upgrades` and (b) such a transaction causes the sources to be
cleared / invalidated until we reverify the sources.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes