[consensus] adding a throughput calculator
## Description 

This is a first part of the "dynamic parameters optimisation" work
proposal that can be [found
here](https://www.notion.so/haneullabs/2023-09-Dynamic-Parameters-Optimisation-2af64149fbe649ed98e3b148aca8f551?pvs=4d829457c4559a53ecac4d8714c71).

This PR is the first steps towards this line of work where a new
component is introduced to calculate the throughput based on the
consensus output. Based on the calculated throughput we do assign a
"traffic profile" which we can later use on downstream on the consensus
submission side so we can make decisions on the transaction
amplification to trade off latency vs throughput.

The introduced `ConsensusThroughputCalculator` is calculating the
throughput by using the uniquely processed transactions from consensus
(after every deduplication attempt) and by using the committed leader's
timestamp for the observations tracking to effectively calculate the
number of transactions to the unit of second. That's a best effort
approach to ensure that validators view are converging as much as
possible.

The `TrafficProfile` is defined based on the lower throughput limit that
corresponds to and currently to keep things simple we consider just two
profiles, `Low` and `High`. A rolling window approach has been chosen to
calculate the throughput based on predefined window size to store the
observations, dictated by the `observations_window` property. Also, the
`traffic_profile_update_window` property is defining some minimum time
that needs to be passed between two traffic profile updates to ensure
that we avoid jitterish behaviour. Finally the
`traffic_profile_cool_down_throughput_threshold` is being used as a way
to define a "cool down" threshold when we are about to change from a
higher traffic profile to a lower one again to ensure that we avoid
flipping between neighbouring traffic profiles when throughput is
constantly varying around the profile throughput limits. An example can
be seen bellow on two runs in an AWS benchmark when considering a `high`
traffic profile for throughput >= 2K tps and `low` when bellow. On the
first run on the `Traffic profile` graph we can see that we constantly
alternate between 0 & 1 (0 = low, 1 = high) , whereas on the second run
where we use the cooldown we see that we are stable at `1 = high` and
only switch to `low` when throughput goes bellow the cool down threshold
(for traffic_profile_cool_down_throughput_threshold = 10% , so 0.9 * 2K
= 1.8K tps).

 
<img width="1709" alt="Screenshot 2023-09-19 at 18 28 09"
src="https://github.com/GeunhwaJeong/haneul/assets/17335598/908e44aa-131d-4000-974b-b39223ffe332">

<img width="1710" alt="Screenshot 2023-09-19 at 19 18 49"
src="https://github.com/GeunhwaJeong/haneul/assets/17335598/1cf62499-31dc-4347-a962-9784fdd798f5">


Next steps:
- [ ] Release in testnet/mainnet to dry run and observe via metrics that
is working robustly in accurately detecting the traffic during the epoch
across all validators
- [ ] Connect to the submission side and make informed decisions on the
amplification according to the current traffic profile (guard behind a
protocol config flag)
- [ ] Keep testing in our dev environments under several conditions
(crashed nodes, change on throughput etc)
- [ ] Tune further the throughput calculator parameters

## Test Plan 

Unit tests & running in AWS test cluster environment via the
`haneul-aws-orchestrator`

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes