Make formal snapshot creation wait until db checkpoint is uploaded
## Description 

Formal snapshot creation starts as soon as it notices a db checkpoint
directory. And tries to open the db to create a live object set
iterator. But because we also do pruning and compaction on the db
checkpoint before we upload, a different thread tries to open the db to
do pruning and compaction at the same time and panics.
The fix is for formal snapshot creation workflow to wait until db
checkpoint is already uploaded before it will consider it.