Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/dapp-kit@0.5.0

### Minor Changes

- 6e2a1dfe0: Add a new createNetworkConfig helper for managing network
specific state

## haneullabs/create-dapp@0.0.5

### Patch Changes

-   Updated dependencies [6e2a1dfe0]
    -   haneullabs/dapp-kit@0.5.0

## haneullabs/sdk-docs@1.0.3

### Patch Changes

-   Updated dependencies [6e2a1dfe0]
    -   haneullabs/dapp-kit@0.5.0