[GraphQL/Schema][RFC] Scalars for Move Types
## Description

Continuing  which converts the Move value representation into a
scalar. This PR is an RFC to do the same for the various type
representations, for similar reasons (they are deeply nested structures
which are difficult to query with GraphQL).

The downside is that we lose the ability to query, e.g. module
information, struct declaration, etc for a type in the same query as the
type (it needs to be issued as a separate query). This seems reasonable,
however, because it is difficult to do this in practice because such a
query would be all or nothing (e.g. the modules of all fields, or all
type parameters, etc).

If the number of scalars is piling too high, we can also return a
generic `JSON` scalar for all of these types (`MoveData`,
`MoveTypeSignature`, `MoveTypeLayout`, `OpenMoveTypeSignature`), but the
type specificity is beneficial for writing clients in typed languages
(like Rust).

## Test Plan

:eyes: