Remove module traits from AuthorityStore
## Description 

AuthorityStore should not care about Move modules. This can be done
because we can retrieve modules as long as a type has
BackingPackageStore trait which could fetch package objects.
Doing so forces us to not depend on the store directly as a module
store.
This allowed me discover a place in RPC/indexerv0 where we forgot to use
the module cache but went to the store directly. Fixed it.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes