[haneul-mode][move-compiler] Fixed init bug. Improved other cases
## Description 

- Fixed bug with OTW checks
- Improved init error messages

## Test Plan 

- Added tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

This release fixes a bug in the Haneul mode of the Move compiler that
categorized certain structs erroneously as one-time witnesses.