[execution] Recompute has_public_transfer on load
## Description 

- has_public_transfer is conceptually static, but is not for core types
due to in-place upgrades
- The flag originally existed to avoid having to load the VM in the very
early days of Haneul. This is no longer the case and the type is loaded
anyway. So we should just recompute the flag.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes