Simplify transactional-test-runner genesis build
## Description 

Simplify how we build genesis in transactional-test-runner.
The genesis building process in the test authority builder already loads
the corresponding packages and objects. We don't need to do them in the
transactional test runner init. They are redundant. Remove all of them.
Also fix a bug in test authority builder where we didn't pass along the
protocol version.
 
## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes