[move][cleanup] Remove global storage operations from source language and IR
## Description 

- Removed the global storage operations from the source language and the
IR
- Docs were started to be cleaned, but this will be more work so there
will still be references to the operations in the docs at this time (the
docs need a bigger rewrite and were already broken)

## Test Plan 

- Updated tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

The Move toolchain no longer supports legacy global storage operations.