[transfer-to-object] Add native errors to module
## Description 

Adds the native error codes for native functions to the transfer module.
Also adds a function to return the object ID of the receiving object.

## Test Plan 

Updated expected value tests -- function handle index changed due to the
new function.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Added `Receiving` native error codes to the `haneul::transfer` module, and
added a function to get the object ID for a `Receiving` argument in
Move.