[GraphQL/RFC] Changes to `asX` fields in Object type hierarchy
## Description

Based on feedback that having to explicitly upcast (e.g. from `Coin`, to
`MoveObject`, to `Object`) was annoying, this PR changes the schema so
that accessing parent type fields is done via interfaces (each parent
type comes with an accompanying interface).

`asX` fields are still used for downcasting, however, e.g. `MoveObject`
has `asCoin`, `asStake`, etc.

This change required adopting a naming scheme for these accompanying
interfaces, e.g. `Foo` is accompanied by `IFoo` (and this was
retroactively applied to `ObjectOwner` which became `IOwner`).

Also cleans up some inconsistencies in the draft schema relative to the
current planned schema:

- Dynamic Field related APIs have moved to the `IOwner` interface, to
make it possible to query dynamic fields on wrapped objects (which can't
otherwise be queried for using the `object` queries currently).
- Added `origins` as a field to `MovePackage` to expose type origin
information (similar to linkage information).
- Added `asCoinMetadata` to `MoveObject` because it's another type that
implements `IMoveObject`.
- Renamed `MovePackage.bcs` to `MovePackage.moduleBcs` as otherwise this
field would clash with `IObject.bcs` (which holds a different value.)
- Have `IOwner.objectConnection` return a `MoveObjectConnection` instead
of an `ObjectConnection`. It's not possible for a package to be owned by
another address, so the `Object`s that come from this field will all be
`MoveObject`s, so we might as well avoid the indirection.

## Test Plan

:eyes: