[cli] Implement Display for BalanceChange type
## Description 

Implements the Display trait for the `BalanceChange` type. This is part
of a larger effort (and multiple PRs) in implementing a nice format
output for `HaneulTransactionBlockResponse` in the CLI, following
laura-makdah's excellent formatting in `HaneulTransactionBlockEffects`.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Implements the `Display` trait for the `BalanceChange` type as part of a larger effort in implementing a good-looking format output for `HaneulTransactionBlockResponse` in the CLI.