[indexerv2] use canonical string with 0x prefix for types
## Description 

1. introduces to_canonical_display -> impl Display for AccountAddress,
TypeTag, and StructTag on the move crate
2. implements to_canonical_string for haneul_types::MoveObjectType
3. test cases that verify Object -> IndexedObject -> StoredObject ->
Object conversion does not lose or result in the incorrect type string/
display
4. test cases to verify that parse_haneul_struct_tag still works correctly

## Test Plan 

unit tests on conversion from haneul_types::Object to IndexedObject to
StoredObject. The dynamic fields path is a bit challenging to write a
unit test for

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes